(ch-database): setup db [Finishes #167418083]
