+++ package.json
@@ -7,13 +7,18 @@
   "license": "MIT",
   "scripts": {
     "test": "nyc --reporter=html --reporter=text mocha --exit './test/**/*.spec.js'",
-    "start:dev": "node index"
+    "start:dev": "node ./express/index",
+    "create:db": "psql fleek -f ./src/database/tables.sql",
+    "drop:db": "psql fleek -f ./src/database/drop.sql"
   },
   "dependencies": {
+    "body-parser": "^1.19.0",
     "chai": "^4.2.0",
+    "dotenv": "^8.0.0",
     "express": "^4.17.1",
     "mocha": "^6.2.0",
-    "nyc": "^14.1.1"
+    "nyc": "^14.1.1",
+    "pg": "^7.11.0"
   },
   "devDependencies": {
     "@babel/core": "^7.5.5",
