+++ src/database/connection.js
@@ -1,0 +1,12 @@
+const Pool = require('pg').Pool
+const { user, host, database, password, port } = require('./config')
+
+const pool = new Pool({
+  user,
+  host,
+  database, 
+  password,
+  port
+})
+
+module.exports = pool
