+++ src/database/drop.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS users;
