+++ src/database/tables.sql
@@ -1,0 +1,9 @@
+CREATE TABLE IF NOT EXISTS people (
+  ID SERIAL PRIMARY KEY,
+  username VARCHAR(50) UNIQUE NOT NULL,
+  email VARCHAR(50) UNIQUE NOT NULL,
+  phone INTEGER UNIQUE NOT NULL,
+  password VARCHAR(50) NOT NULL,
+  created_on TIMESTAMP NOT NULL,
+  updated_on TIMESTAMP NOT NULL
+)
