+++ test/index.spec.js
@@ -1,5 +1,5 @@
 const expect = require('chai').expect 
-const server = require('../index')
+const server = require('../express/index')
 
 describe('test', () => {
   it('should return a string', () => {
